requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
elasticsearch==7.12.0
et-xmlfile==1.0.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.3
pathspec==0.8.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
regex==2021.3.17
requests==2.25.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
